requirements.txt
django==5.0.8
djangorestframework==3.15.2
django-cors-headers==4.4.0
django-unfold==0.19.0
environs==11.0.0
gunicorn==23.0.0